Course 4/requirements.txt
opencv-python>=4.10.0
numpy>=1.26.0
Pillow>=10.0.0
ultralytics>=8.3.0
